requirements.txt
Pillow
pandas
scikit-learn
numpy
keras
tensorflow
matplotlib
